Update Open LLM Leaderboard source urls
Related to https://huggingface.co/spaces/HuggingFaceH4/open_llm_leaderboard/discussions/462.

It would be nice to allow deeplinks in the Open LLM Leaderboard to load the leaderboard on a specific model. Once that's the case, we should share those as source url in the model card metadata (which is what this PR do). 

cc @Weyaxi if you want to integrate this in your leaderboard bot :)
(and cc @clefourrier @julien-c for viz')